amoutn now is float

+++ QuetzalBank/Models/Transferences/MovementModel.swift
@@ -11,7 +11,7 @@ struct MovementModel: Codable {
     let id: Int
     let id_sender: Int
     let id_receptor: Int
-    let amount: Int
+    let amount: Float
     let sender_account: String
     let receptor_account: String
 }
